Move all prop-types to types folder
