+++ project/src/types/location.prop.js
@@ -1,0 +1,7 @@
+import PropTypes from 'prop-types';
+
+export default PropTypes.shape({
+  latitude: PropTypes.number.isRequired,
+  longitude: PropTypes.number.isRequired,
+  zoom: PropTypes.number.isRequired,
+});
