+++ project/src/types/offers.prop.js
@@ -1,4 +1,4 @@
 import PropTypes from 'prop-types';
-import offersProp from '../offer/offer.prop';
+import offersProp from './offer.prop';
 
 export default PropTypes.arrayOf(offersProp);
